example: removing utils.rs
